refactor: ada perubahan pada route admin

+++ routes/admin.php
@@ -1,31 +1,34 @@
 <?php
 
 use App\Http\Controllers\KerusakanController;
+use App\Http\Controllers\RuleController;
+use App\Http\Controllers\PertanyaanController;
+use App\Http\Controllers\GejalaController;
 use Illuminate\Support\Facades\Route;
 
 
-Route::middleware(['auth', 'isAdmin'])->group(function () {
+Route::prefix('admin')->middleware(['auth', 'isAdmin'])->group(function () {
 
     Route::get('/dashboard', function () {
         return view('pages.admin.dashboard');
-    })->name('dashboard');
+    })->name('admin.dashboard');
 
     Route::prefix('kerusakan')->group(function(){
-        Route::get('/', [KerusakanController::class, 'index'])->name('kerusakan.index');
-        Route::get('/create', [KerusakanController::class, 'create'])->name('kerusakan.create');
-        Route::post('/store', [KerusakanController::class, 'store'])->name('kerusakan.store');
-        Route::get('/edit/{id}', [KerusakanController::class, 'edit'])->name('kerusakan.edit');
-        Route::put('/update/{id}', [KerusakanController::class, 'update'])->name('kerusakan.update');
-        Route::delete('/destroy/{id}', [KerusakanController::class, 'destroy'])->name('kerusakan.destroy');
+        Route::get('/', [KerusakanController::class, 'index'])->name('admin.kerusakan.index');
+        Route::get('/create', [KerusakanController::class, 'create'])->name('admin.kerusakan.create');
+        Route::post('/store', [KerusakanController::class, 'store'])->name('admin.kerusakan.store');
+        Route::get('/edit/{id}', [KerusakanController::class, 'edit'])->name('admin.kerusakan.edit');
+        Route::put('/update/{id}', [KerusakanController::class, 'update'])->name('admin.kerusakan.update');
+        Route::delete('/destroy/{id}', [KerusakanController::class, 'destroy'])->name('admin.kerusakan.destroy');
     });
 
     Route::prefix('gejala')->group(function(){
-        Route::get('/', [App\Http\Controllers\GejalaController::class, 'index'])->name('gejala.index');
-        Route::get('/create', [App\Http\Controllers\GejalaController::class, 'create'])->name('gejala.create');
-        Route::post('/store', [App\Http\Controllers\GejalaController::class, 'store'])->name('gejala.store');
-        Route::get('/edit/{id}', [App\Http\Controllers\GejalaController::class, 'edit'])->name('gejala.edit');
-        Route::put('/update/{id}', [App\Http\Controllers\GejalaController::class, 'update'])->name('gejala.update');
-        Route::delete('/destroy/{id}', [App\Http\Controllers\GejalaController::class, 'destroy'])->name('gejala.destroy');
+        Route::get('/', [GejalaController::class, 'index'])->name('admin.gejala.index');
+        Route::get('/create', [GejalaController::class, 'create'])->name('admin.gejala.create');
+        Route::post('/store', [GejalaController::class, 'store'])->name('admin.gejala.store');
+        Route::get('/edit/{id}', [GejalaController::class, 'edit'])->name('admin.gejala.edit');
+        Route::put('/update/{id}', [GejalaController::class, 'update'])->name('admin.gejala.update');
+        Route::delete('/destroy/{id}', [GejalaController::class, 'destroy'])->name('admin.gejala.destroy');
     });
